hybrid-ssr-ocl-full-extended/requirements.txt
z3-solver>=4.13
pydantic>=2.7
typer>=0.12
rich>=13.7
pyyaml>=6.0
jinja2>=3.1
pytest>=8.0
flake8>=7.0
# Neural-Symbolic components
torch>=2.0.0
transformers>=4.21.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
